feat: ステップ全体で扱う値をプロバイダーに分割
