+++ src/app/ar-assets/create/page.tsx
@@ -1,4 +1,5 @@
 import { CreateArAssetStepper } from './_components/CreateArAssetStepper';
+import { RequestBodiesProvider } from './_components/RequestBodiesProvider';
 import { Guard } from '@/shared/components/features';
 import { GlobalNav } from '@/shared/components/layouts/GlobalNav';
 
@@ -6,7 +7,9 @@ const Page = () => {
   return (
     <Guard>
       <GlobalNav>
-        <CreateArAssetStepper />
+        <RequestBodiesProvider>
+          <CreateArAssetStepper />
+        </RequestBodiesProvider>
       </GlobalNav>
     </Guard>
   );
